moving the status bar to the widget
